create commands: create delete and update for selfassessmentsValue

+++ src/Application/Journals/Command/DeleteJournal/DeleteJournalCommandHandler.cs
@@ -15,9 +15,11 @@
         {
             var journal = await _journalRepository.GetByIdAsync(request.JournalId, cancellationToken);
 
-            _journalRepository.Remove(journal);
-
-            await _unitOfWork.SaveChangesAsync(cancellationToken);
+            if (journal == null)
+            {
+                _journalRepository.Remove(journal);
+                await _unitOfWork.SaveChangesAsync(cancellationToken);
+            }
 
             return Unit.Value;
         }
